+++ resolwe/flow/models/entity.py
@@ -11,6 +11,7 @@ from resolwe.flow.models.annotations import (
     AnnotationValue,
     HandleMissingAnnotations,
 )
+from resolwe.observers.consumers import BackgroundTaskType
 from resolwe.observers.decorators import move_to_container
 from resolwe.observers.models import BackgroundTask
 from resolwe.observers.utils import start_background_task
@@ -24,23 +25,18 @@ from .utils import DirtyError, validate_schema
 class EntityQuerySet(BaseQuerySet, PermissionQuerySet):
     """Query set for ``Entity`` objects."""
 
-    def duplicate(self, contributor, inherit_entity=False, inherit_collection=False):
-        """Duplicate (make a copy) ``Entity`` objects in the background.
-
-        :param contributor: Duplication user
-        """
-        task = BackgroundTask.objects.create(description="Duplicate entity")
-        start_background_task(
-            {
-                "type": "duplicate_entity",
-                "entity_ids": list(self.values_list("pk", flat=True)),
-                "task_id": task.id,
-                "contributor_id": contributor.id,
-                "inherit_entity": inherit_entity,
-                "inherit_collection": inherit_collection,
-            }
+    def duplicate(self, contributor, inherit_collection=False) -> BackgroundTask:
+        """Duplicate (make a copy) ``Entity`` objects in the background."""
+        task_data = {
+            "entity_ids": list(self.values_list("pk", flat=True)),
+            "inherit_collection": inherit_collection,
+        }
+        return start_background_task(
+            BackgroundTaskType.DUPLICATE_ENTITY,
+            "Duplicate entity",
+            task_data,
+            contributor,
         )
-        return task
 
     @transaction.atomic
     def move_to_collection(
@@ -225,22 +221,15 @@ class Entity(BaseCollection, PermissionObject):
         """Return True if entity is a duplicate."""
         return bool(self.duplicated)
 
-    def duplicate(self, contributor, inherit_collection=False):
-        """Duplicate (make a copy) ``Data`` objects in the background.
-
-        :param contributor: Duplication user
-        """
-        task = BackgroundTask.objects.create(description="Duplicate entity")
-        start_background_task(
-            {
-                "type": "duplicate_entity",
-                "entity_ids": [self.pk],
-                "task_id": task.id,
-                "contributor_id": contributor.id,
-                "inherit_collection": inherit_collection,
-            }
+    def duplicate(self, contributor, inherit_collection=False) -> BackgroundTask:
+        """Duplicate (make a copy) object in the background."""
+        task_data = {"entity_ids": [self.pk], "inherit_collection": inherit_collection}
+        return start_background_task(
+            BackgroundTaskType.DUPLICATE_ENTITY,
+            "Duplicate entity",
+            task_data,
+            contributor,
         )
-        return task
 
     @transaction.atomic
     @move_to_container